+++ R/simonEfficacy.R
@@ -46,7 +46,6 @@ simonEfficacy <- function(n1, n2, r1, r, e1, p0, p1)
   n.curt.s1 <- rep(n1, length(k.curt.s1))
   curtail.s1 <- cbind(k.curt.s1, n.curt.s1, prob.curt.s1, prob.curt.s1.p0)
 
-
   ############## S2 ###############
 
   # Pick out the coefficients for the S2 paths (A, say):
@@ -73,23 +72,19 @@ simonEfficacy <- function(n1, n2, r1, r, e1, p0, p1)
 
   go <- cbind(k.go, n.go, prob.go, prob.go.p0)
 
-  final <- rbind(curtail.s1, go)
-
+  output <- rbind(curtail.s1, go)
+  colnames(output) <- c("k", "n", "prob", "prob.p0")
 
   ############## WRAPPING UP THE RESULTS ##############
+  #output <- cbind(output, rep(0, nrow(output)))
 
-  output <- data.frame(k=final[,1], n=final[,2], prob=final[,3], prob.p0=final[,4])
-
-  output$success <- "Fail"
-  output$success[output$k > r] <- "Success"
-  output$success[output$n==n1 & output$k > e1] <- "Success"
-
+  success <- output[, "k"] > r | ( output[, "n"]==n1 & output[, "k"] > e1)
 
   # Pr(early termination):
   #PET <- sum(output$prob[output$n < n])
   #PET.p0 <- sum(output$prob.p0[output$n < n])
 
-  power <- sum(output$prob[output$success=="Success"])
+  power <- sum(output[, "prob"][success==TRUE])
 
   #output$obsd.p <- output$k/output$n
 
@@ -100,14 +95,14 @@ simonEfficacy <- function(n1, n2, r1, r, e1, p0, p1)
 
   #sample.size <- wtd.quantile(output$n, weights=output$prob, normwt=TRUE, probs=c(0.25, 0.5, 0.75))
   #sample.size.expd <- wtd.mean(output$n, weights=output$prob, normwt=TRUE)
-  sample.size.expd <- sum(output$n*output$prob)
+  sample.size.expd <- sum(output[,"n"]*output[, "prob"])
 
   #sample.size.p0 <- wtd.quantile(output$n, weights=output$prob.p0, normwt=TRUE, probs=c(0.25, 0.5, 0.75))
   #sample.size.expd.p0 <- wtd.mean(output$n, weights=output$prob.p0, normwt=TRUE)
-  sample.size.expd.p0 <- sum(output$n*output$prob.p0)
+  sample.size.expd.p0 <- sum(output[,"n"]*output[, "prob.p0"])
 
 
-  alpha <- sum(output$prob.p0[output$success=="Success"])
+  alpha <- sum(output[, "prob.p0"][success==TRUE])
 
   #output <- list(output, mean.bias=bias.mean, var.bias=bias.var, sample.size=sample.size, expd.sample.size=sample.size.expd, PET=PET,
   #               sample.size.p0=sample.size.p0, expd.sample.size.p0=sample.size.expd.p0, PET.p0=PET.p0, alpha=alpha, power=power)